Add simple searching of all question fields
Why is this change needed?
--------------------------
As the number of questions answered increase, it becomes more and more
difficult to find one that is relevant to you, or even one you know
exists.

How does it address the issue?
------------------------------
This is a naive implementation of full text search for questions. There
may be a way to do it but I couldn't figure out how to do substring
matching on tags so if you want to search by tag you must provide the
full tag name.

This also doesn't do any ranking or highlighting of matches. I suspect a
different approach that would work well would be adding all the question
metadata into the page and doing JS searches, though that breaks down
once you start paginating.

Also this is "naive" because it doesn't actually use any of postgres's
full text features.

Any links to any relevant tickets, articles or other resources?
--------------------------------------------------------------
Some helpful links for this version were:
* [Fast Full-text Search with Ecto and
PostgreSQL](https://nathanmlong.com/2018/01/fast-fulltext-search-with-ecto-and-postgresql/)
* [Postgres full-text search is Good
Enough!](http://rachbelaid.com/postgres-full-text-search-is-good-enough/)

I used neither of them for the solution, but many further improvements
could be gotten by following along what they did.

Issues Addressed
---------------
* Fixes #4
* Fixes #5
* Fixes #6
* Fixes #7